Model predictions from fitted fit_ageclass objects

Model predictions from fitted fit_ageclass objects.

Usage

# S3 method for class 's4t_ageclass_model'
predict(object, newdata, type = c("prob", "class", "cum.prob"), ...)

Arguments

object

a s4t_ageclass_model object

newdata

A data frame object in the same format as aux_age_df (see documentation for s4t_ch) containing data to predict age-class. If omitted, the fitted values for the aux_age_df data in the s4t_ch from the fitted model object are returned.

type

The prediction type. A character of one of the three option prob, class, or cum.prob.

...

not used (needed for generic consistency)

Value

Depends on the type. type = "prob" returns a matrix of the probabilities, type = "class" returns a vector of the mostly likely age-class (highest probability), type = "cum.prob" returns the cumulative probabilities as a matrix (probability that the individual is at least a particular age-class).

Examples

sim.dat <- sim_simple_s4t_ch(N = 2000)
m1 <- fit_ageclass(age_formula = ~ FL,
                   s4t_ch = sim.dat$s4t_ch)
predict(m1,type = "prob")
#>                 Age1       Age2         Age3
#>    [1,] 0.0006633707 0.03347510 0.9658615264
#>    [2,] 0.0065094031 0.25212865 0.7413619475
#>    [3,] 0.1218248156 0.75893608 0.1192391021
#>    [4,] 0.0007297230 0.03669798 0.9625723009
#>    [5,] 0.9560936488 0.04304463 0.0008617232
#>    [6,] 0.6881019791 0.30345703 0.0084409951
#>    [7,] 0.6010090895 0.38667643 0.0123144757
#>    [8,] 0.0865227271 0.74800552 0.1654717513
#>    [9,] 0.9722879205 0.02717708 0.0005350041
#>   [10,] 0.0022890024 0.10657185 0.8911391452
#>   [11,] 0.6236497672 0.36514367 0.0112065588
#>   [12,] 0.3240773045 0.63822831 0.0376943886
#>   [13,] 0.4125321738 0.56141955 0.0260482749
#>   [14,] 0.9519101514 0.04714195 0.0009478964
#>   [15,] 0.1826828923 0.73980515 0.0775119616
#>   [16,] 0.1324044166 0.75801699 0.1095785945
#>   [17,] 0.0865227271 0.74800552 0.1654717513
#>   [18,] 0.9473499917 0.05160733 0.0010426780
#>   [19,] 0.8739152198 0.12338248 0.0027022993
#>   [20,] 0.6881019791 0.30345703 0.0084409951
#>   [21,] 0.0292661881 0.58689787 0.3838359381
#>   [22,] 0.0095049939 0.32866277 0.6618322406
#>   [23,] 0.9473499917 0.05160733 0.0010426780
#>   [24,] 0.3453145175 0.62030292 0.0343825630
#>   [25,] 0.0033489355 0.14841397 0.8482370954
#>   [26,] 0.0078669861 0.28899942 0.7031335946
#>   [27,] 0.0009712804 0.04824762 0.9507810980
#>   [28,] 0.0321015738 0.60636051 0.3615379128
#>   [29,] 0.0008829833 0.04405884 0.9550581783
#>   [30,] 0.7804425949 0.21430167 0.0052557396
#>   [31,] 0.0020811634 0.09786459 0.9000542423
#>   [32,] 0.0027688435 0.12602839 0.8712027653
#>   [33,] 0.9178144325 0.08050666 0.0016789053
#>   [34,] 0.3671874062 0.60146036 0.0313522340
#>   [35,] 0.6236497672 0.36514367 0.0112065588
#>   [36,] 0.1028402893 0.75638384 0.1407758676
#>   [37,] 0.0943660319 0.75291894 0.1527150314
#>   [38,] 0.1688678048 0.74651810 0.0846141002
#>   [39,] 0.3240773045 0.63822831 0.0376943886
#>   [40,] 0.2647764639 0.68565825 0.0495652829
#>   [41,] 0.0865227271 0.74800552 0.1654717513
#>   [42,] 0.1028402893 0.75638384 0.1407758676
#>   [43,] 0.0126147257 0.39223982 0.5951454579
#>   [44,] 0.1826828923 0.73980515 0.0775119616
#>   [45,] 0.1119814537 0.75839125 0.1296272985
#>   [46,] 0.9695983855 0.02981309 0.0005885243
#>   [47,] 0.6010090895 0.38667643 0.0123144757
#>   [48,] 0.2129087263 0.72216887 0.0649223996
#>   [49,] 0.9423835842 0.05646949 0.0011469260
#>   [50,] 0.0126147257 0.39223982 0.5951454579
#>   [51,] 0.0555227860 0.70235498 0.2421222350
#>   [52,] 0.0027688435 0.12602839 0.8712027653
#>   [53,] 0.8739152198 0.12338248 0.0027022993
#>   [54,] 0.0507271306 0.68921775 0.2600551230
#>   [55,] 0.8739152198 0.12338248 0.0027022993
#>   [56,] 0.8934830755 0.10428296 0.0022339646
#>   [57,] 0.0030451493 0.13684074 0.8601141119
#>   [58,] 0.9634412713 0.03584658 0.0007121504
#>   [59,] 0.0183680063 0.48070983 0.5009221599
#>   [60,] 0.0114801098 0.37061256 0.6179073322
#>   [61,] 0.0006633707 0.03347510 0.9658615264
#>   [62,] 0.9247294920 0.07374413 0.0015263777
#>   [63,] 0.0014219058 0.06905302 0.9295250721
#>   [64,] 0.0266743374 0.56669234 0.4066333204
#>   [65,] 0.1558981819 0.75179995 0.0923018700
#>   [66,] 0.9666567752 0.03269583 0.0006473950
#>   [67,] 0.1218248156 0.75893608 0.1192391021
#>   [68,] 0.1826828923 0.73980515 0.0775119616
#>   [69,] 0.0022890024 0.10657185 0.8911391452
#>   [70,] 0.9473499917 0.05160733 0.0010426780
#>   [71,] 0.0095049939 0.32866277 0.6618322406
#>   [72,] 0.0352016977 0.62497746 0.3398208386
#>   [73,] 0.9695983855 0.02981309 0.0005885243
#>   [74,] 0.0059206301 0.23485046 0.7592289137
#>   [75,] 0.6236497672 0.36514367 0.0112065588
#>   [76,] 0.0266743374 0.56669234 0.4066333204
#>   [77,] 0.8630233756 0.13400464 0.0029719845
#>   [78,] 0.0010683977 0.05281167 0.9461199348
#>   [79,] 0.0507271306 0.68921775 0.2600551230
#>   [80,] 0.2837590220 0.67098144 0.0452595352
#>   [81,] 0.0243062768 0.54585357 0.4298401487
#>   [82,] 0.0071563047 0.27019024 0.7226534590
#>   [83,] 0.0015640102 0.07542209 0.9230139015
#>   [84,] 0.6672700055 0.32345193 0.0092780648
#>   [85,] 0.4358283827 0.54043718 0.0237344388
#>   [86,] 0.0114801098 0.37061256 0.6179073322
#>   [87,] 0.8934830755 0.10428296 0.0022339646
#>   [88,] 0.0555227860 0.70235498 0.2421222350
#>   [89,] 0.9423835842 0.05646949 0.0011469260
#>   [90,] 0.0022890024 0.10657185 0.8911391452
#>   [91,] 0.0015640102 0.07542209 0.9230139015
#>   [92,] 0.9695983855 0.02981309 0.0005885243
#>   [93,] 0.0010683977 0.05281167 0.9461199348
#>   [94,] 0.0321015738 0.60636051 0.3615379128
#>   [95,] 0.0607428037 0.71420734 0.2250498540
#>   [96,] 0.9695983855 0.02981309 0.0005885243
#>   [97,] 0.7460116445 0.24763482 0.0063535341
#>   [98,] 0.3671874062 0.60146036 0.0313522340
#>   [99,] 0.0020811634 0.09786459 0.9000542423
#>  [100,] 0.9423835842 0.05646949 0.0011469260
#>  [101,] 0.0006633707 0.03347510 0.9658615264
#>  [102,] 0.3453145175 0.62030292 0.0343825630
#>  [103,] 0.6881019791 0.30345703 0.0084409951
#>  [104,] 0.0221436621 0.52449686 0.4533594812
#>  [105,] 0.8739152198 0.12338248 0.0027022993
#>  [106,] 0.9473499917 0.05160733 0.0010426780
#>  [107,] 0.0059206301 0.23485046 0.7592289137
#>  [108,] 0.1973598138 0.73168045 0.0709597348
#>  [109,] 0.0040500672 0.17393688 0.8220130518
#>  [110,] 0.0138599071 0.41417318 0.5719669157
#>  [111,] 0.8934830755 0.10428296 0.0022339646
#>  [112,] 0.8934830755 0.10428296 0.0022339646
#>  [113,] 0.0138599071 0.41417318 0.5719669157
#>  [114,] 0.9747457180 0.02476793 0.0004863486
#>  [115,] 0.0044536567 0.18792174 0.8076246057
#>  [116,] 0.9423835842 0.05646949 0.0011469260
#>  [117,] 0.0463253724 0.67484706 0.2788275631
#>  [118,] 0.0048972662 0.20273593 0.7923668033
#>  [119,] 0.8113861288 0.18426707 0.0043467976
#>  [120,] 0.3035406329 0.65514780 0.0413115708
#>  [121,] 0.0044536567 0.18792174 0.8076246057
#>  [122,] 0.0664190731 0.72473139 0.2088495383
#>  [123,] 0.2466265168 0.69911611 0.0542573773
#>  [124,] 0.6010090895 0.38667643 0.0123144757
#>  [125,] 0.7275158721 0.26549907 0.0069850545
#>  [126,] 0.9599285299 0.03928809 0.0007833779
#>  [127,] 0.0036829155 0.16077261 0.8355444721
#>  [128,] 0.9560936488 0.04304463 0.0008617232
#>  [129,] 0.0040500672 0.17393688 0.8220130518
#>  [130,] 0.0009712804 0.04824762 0.9507810980
#>  [131,] 0.7275158721 0.26549907 0.0069850545
#>  [132,] 0.0266743374 0.56669234 0.4066333204
#>  [133,] 0.0555227860 0.70235498 0.2421222350
#>  [134,] 0.2647764639 0.68565825 0.0495652829
#>  [135,] 0.0004983522 0.02536342 0.9741382322
#>  [136,] 0.8513506805 0.14538082 0.0032684955
#>  [137,] 0.9473499917 0.05160733 0.0010426780
#>  [138,] 0.2466265168 0.69911611 0.0542573773
#>  [139,] 0.9022272784 0.09574161 0.0020311127
#>  [140,] 0.0005482079 0.02782876 0.9716230340
#>  [141,] 0.0664190731 0.72473139 0.2088495383
#>  [142,] 0.8255540056 0.17049314 0.0039528532
#>  [143,] 0.1119814537 0.75839125 0.1296272985
#>  [144,] 0.0065094031 0.25212865 0.7413619475
#>  [145,] 0.9769906910 0.02256719 0.0004421161
#>  [146,] 0.3453145175 0.62030292 0.0343825630
#>  [147,] 0.0004530285 0.02311107 0.9764359019
#>  [148,] 0.6236497672 0.36514367 0.0112065588
#>  [149,] 0.2293326388 0.71130146 0.0593659045
#>  [150,] 0.0725847849 0.73389103 0.1935241839
#>  [151,] 0.3453145175 0.62030292 0.0343825630
#>  [152,] 0.0011752142 0.05778007 0.9410447184
#>  [153,] 0.1437523831 0.75563626 0.1006113548
#>  [154,] 0.3240773045 0.63822831 0.0376943886
#>  [155,] 0.4125321738 0.56141955 0.0260482749
#>  [156,] 0.0138599071 0.41417318 0.5719669157
#>  [157,] 0.2647764639 0.68565825 0.0495652829
#>  [158,] 0.0792742605 0.74165687 0.1790688667
#>  [159,] 0.0385892715 0.64265374 0.3187569901
#>  [160,] 0.9473499917 0.05160733 0.0010426780
#>  [161,] 0.0036829155 0.16077261 0.8355444721
#>  [162,] 0.7636598029 0.23056142 0.0057787774
#>  [163,] 0.1028402893 0.75638384 0.1407758676
#>  [164,] 0.5545017753 0.43063359 0.0148646340
#>  [165,] 0.2293326388 0.71130146 0.0593659045
#>  [166,] 0.0725847849 0.73389103 0.1935241839
#>  [167,] 0.1437523831 0.75563626 0.1006113548
#>  [168,] 0.4831775351 0.49712946 0.0196930065
#>  [169,] 0.6672700055 0.32345193 0.0092780648
#>  [170,] 0.0033489355 0.14841397 0.8482370954
#>  [171,] 0.9747457180 0.02476793 0.0004863486
#>  [172,] 0.0086476293 0.30850952 0.6828428487
#>  [173,] 0.0086476293 0.30850952 0.6828428487
#>  [174,] 0.1437523831 0.75563626 0.1006113548
#>  [175,] 0.0167246838 0.45852407 0.5247512480
#>  [176,] 0.9022272784 0.09574161 0.0020311127
#>  [177,] 0.9423835842 0.05646949 0.0011469260
#>  [178,] 0.0607428037 0.71420734 0.2250498540
#>  [179,] 0.9560936488 0.04304463 0.0008617232
#>  [180,] 0.0943660319 0.75291894 0.1527150314
#>  [181,] 0.0078669861 0.28899942 0.7031335946
#>  [182,] 0.4594113593 0.51896707 0.0216215750
#>  [183,] 0.0017202918 0.08232415 0.9159555561
#>  [184,] 0.0009712804 0.04824762 0.9507810980
#>  [185,] 0.2647764639 0.68565825 0.0495652829
#>  [186,] 0.0030451493 0.13684074 0.8601141119
#>  [187,] 0.0352016977 0.62497746 0.3398208386
#>  [188,] 0.2837590220 0.67098144 0.0452595352
#>  [189,] 0.9311063951 0.06750592 0.0013876879
#>  [190,] 0.0104464658 0.34939031 0.6401632282
#>  [191,] 0.9247294920 0.07374413 0.0015263777
#>  [192,] 0.1688678048 0.74651810 0.0846141002
#>  [193,] 0.0004983522 0.02536342 0.9741382322
#>  [194,] 0.0664190731 0.72473139 0.2088495383
#>  [195,] 0.6010090895 0.38667643 0.0123144757
#>  [196,] 0.0321015738 0.60636051 0.3615379128
#>  [197,] 0.0422885536 0.65930283 0.2984086205
#>  [198,] 0.0005482079 0.02782876 0.9716230340
#>  [199,] 0.0006030481 0.03052592 0.9688710360
#>  [200,] 0.0292661881 0.58689787 0.3838359381
#>  [201,] 0.9560936488 0.04304463 0.0008617232
#>  [202,] 0.0792742605 0.74165687 0.1790688667
#>  [203,] 0.0059206301 0.23485046 0.7592289137
#>  [204,] 0.3896212343 0.58179769 0.0285810797
#>  [205,] 0.1324044166 0.75801699 0.1095785945
#>  [206,] 0.0555227860 0.70235498 0.2421222350
#>  [207,] 0.8113861288 0.18426707 0.0043467976
#>  [208,] 0.0138599071 0.41417318 0.5719669157
#>  [209,] 0.0865227271 0.74800552 0.1654717513
#>  [210,] 0.9022272784 0.09574161 0.0020311127
#>  [211,] 0.9695983855 0.02981309 0.0005885243
#>  [212,] 0.6881019791 0.30345703 0.0084409951
#>  [213,] 0.9473499917 0.05160733 0.0010426780
#>  [214,] 0.4594113593 0.51896707 0.0216215750
#>  [215,] 0.0008829833 0.04405884 0.9550581783
#>  [216,] 0.3896212343 0.58179769 0.0285810797
#>  [217,] 0.0018921601 0.08979348 0.9083143594
#>  [218,] 0.0086476293 0.30850952 0.6828428487
#>  [219,] 0.0167246838 0.45852407 0.5247512480
#>  [220,] 0.0044536567 0.18792174 0.8076246057
#>  [221,] 0.0044536567 0.18792174 0.8076246057
#>  [222,] 0.9178144325 0.08050666 0.0016789053
#>  [223,] 0.9311063951 0.06750592 0.0013876879
#>  [224,] 0.1119814537 0.75839125 0.1296272985
#>  [225,] 0.8255540056 0.17049314 0.0039528532
#>  [226,] 0.0321015738 0.60636051 0.3615379128
#>  [227,] 0.3240773045 0.63822831 0.0376943886
#>  [228,] 0.0025175453 0.11594896 0.8815334953
#>  [229,] 0.2837590220 0.67098144 0.0452595352
#>  [230,] 0.0053848227 0.21838102 0.7762341525
#>  [231,] 0.0025175453 0.11594896 0.8815334953
#>  [232,] 0.4358283827 0.54043718 0.0237344388
#>  [233,] 0.0025175453 0.11594896 0.8815334953
#>  [234,] 0.0022890024 0.10657185 0.8911391452
#>  [235,] 0.9599285299 0.03928809 0.0007833779
#>  [236,] 0.4358283827 0.54043718 0.0237344388
#>  [237,] 0.9519101514 0.04714195 0.0009478964
#>  [238,] 0.0030451493 0.13684074 0.8601141119
#>  [239,] 0.5779281724 0.40854140 0.0135304259
#>  [240,] 0.1218248156 0.75893608 0.1192391021
#>  [241,] 0.0025175453 0.11594896 0.8815334953
#>  [242,] 0.0014219058 0.06905302 0.9295250721
#>  [243,] 0.0006030481 0.03052592 0.9688710360
#>  [244,] 0.0095049939 0.32866277 0.6618322406
#>  [245,] 0.2647764639 0.68565825 0.0495652829
#>  [246,] 0.0053848227 0.21838102 0.7762341525
#>  [247,] 0.5308306887 0.45284108 0.0163282280
#>  [248,] 0.0005482079 0.02782876 0.9716230340
#>  [249,] 0.1437523831 0.75563626 0.1006113548
#>  [250,] 0.9103256959 0.08782766 0.0018466465
#>  [251,] 0.9178144325 0.08050666 0.0016789053
#>  [252,] 0.0025175453 0.11594896 0.8815334953
#>  [253,] 0.5308306887 0.45284108 0.0163282280
#>  [254,] 0.9599285299 0.03928809 0.0007833779
#>  [255,] 0.4594113593 0.51896707 0.0216215750
#>  [256,] 0.0065094031 0.25212865 0.7413619475
#>  [257,] 0.0607428037 0.71420734 0.2250498540
#>  [258,] 0.0044536567 0.18792174 0.8076246057
#>  [259,] 0.0036829155 0.16077261 0.8355444721
#>  [260,] 0.7963514349 0.19886875 0.0047798144
#>  [261,] 0.0007297230 0.03669798 0.9625723009
#>  [262,] 0.0027688435 0.12602839 0.8712027653
#>  [263,] 0.7460116445 0.24763482 0.0063535341
#>  [264,] 0.8255540056 0.17049314 0.0039528532
#>  [265,] 0.0221436621 0.52449686 0.4533594812
#>  [266,] 0.6457629349 0.34403977 0.0101972907
#>  [267,] 0.8630233756 0.13400464 0.0029719845
#>  [268,] 0.9178144325 0.08050666 0.0016789053
#>  [269,] 0.0053848227 0.21838102 0.7762341525
#>  [270,] 0.9247294920 0.07374413 0.0015263777
#>  [271,] 0.8255540056 0.17049314 0.0039528532
#>  [272,] 0.2647764639 0.68565825 0.0495652829
#>  [273,] 0.6236497672 0.36514367 0.0112065588
#>  [274,] 0.0020811634 0.09786459 0.9000542423
#>  [275,] 0.9311063951 0.06750592 0.0013876879
#>  [276,] 0.6010090895 0.38667643 0.0123144757
#>  [277,] 0.0292661881 0.58689787 0.3838359381
#>  [278,] 0.9473499917 0.05160733 0.0010426780
#>  [279,] 0.0009712804 0.04824762 0.9507810980
#>  [280,] 0.0022890024 0.10657185 0.8911391452
#>  [281,] 0.0221436621 0.52449686 0.4533594812
#>  [282,] 0.7963514349 0.19886875 0.0047798144
#>  [283,] 0.7963514349 0.19886875 0.0047798144
#>  [284,] 0.0607428037 0.71420734 0.2250498540
#>  [285,] 0.0266743374 0.56669234 0.4066333204
#>  [286,] 0.0152261032 0.43630574 0.5484681554
#>  [287,] 0.7963514349 0.19886875 0.0047798144
#>  [288,] 0.0071563047 0.27019024 0.7226534590
#>  [289,] 0.0792742605 0.74165687 0.1790688667
#>  [290,] 0.6457629349 0.34403977 0.0101972907
#>  [291,] 0.0943660319 0.75291894 0.1527150314
#>  [292,] 0.0011752142 0.05778007 0.9410447184
#>  [293,] 0.4831775351 0.49712946 0.0196930065
#>  [294,] 0.7460116445 0.24763482 0.0063535341
#>  [295,] 0.0020811634 0.09786459 0.9000542423
#>  [296,] 0.0183680063 0.48070983 0.5009221599
#>  [297,] 0.1826828923 0.73980515 0.0775119616
#>  [298,] 0.0086476293 0.30850952 0.6828428487
#>  [299,] 0.8739152198 0.12338248 0.0027022993
#>  [300,] 0.9560936488 0.04304463 0.0008617232
#>  [301,] 0.7804425949 0.21430167 0.0052557396
#>  [302,] 0.0943660319 0.75291894 0.1527150314
#>  [303,] 0.0017202918 0.08232415 0.9159555561
#>  [304,] 0.8934830755 0.10428296 0.0022339646
#>  [305,] 0.5308306887 0.45284108 0.0163282280
#>  [306,] 0.0065094031 0.25212865 0.7413619475
#>  [307,] 0.0221436621 0.52449686 0.4533594812
#>  [308,] 0.0027688435 0.12602839 0.8712027653
#>  [309,] 0.3896212343 0.58179769 0.0285810797
#>  [310,] 0.3896212343 0.58179769 0.0285810797
#>  [311,] 0.0025175453 0.11594896 0.8815334953
#>  [312,] 0.2129087263 0.72216887 0.0649223996
#>  [313,] 0.9747457180 0.02476793 0.0004863486
#>  [314,] 0.0036829155 0.16077261 0.8355444721
#>  [315,] 0.9103256959 0.08782766 0.0018466465
#>  [316,] 0.0005482079 0.02782876 0.9716230340
#>  [317,] 0.9103256959 0.08782766 0.0018466465
#>  [318,] 0.0010683977 0.05281167 0.9461199348
#>  [319,] 0.6010090895 0.38667643 0.0123144757
#>  [320,] 0.0012926963 0.06318328 0.9355240221
#>  [321,] 0.0040500672 0.17393688 0.8220130518
#>  [322,] 0.0044536567 0.18792174 0.8076246057
#>  [323,] 0.0033489355 0.14841397 0.8482370954
#>  [324,] 0.0030451493 0.13684074 0.8601141119
#>  [325,] 0.0321015738 0.60636051 0.3615379128
#>  [326,] 0.7460116445 0.24763482 0.0063535341
#>  [327,] 0.0011752142 0.05778007 0.9410447184
#>  [328,] 0.0167246838 0.45852407 0.5247512480
#>  [329,] 0.0033489355 0.14841397 0.8482370954
#>  [330,] 0.5545017753 0.43063359 0.0148646340
#>  [331,] 0.6881019791 0.30345703 0.0084409951
#>  [332,] 0.0086476293 0.30850952 0.6828428487
#>  [333,] 0.6457629349 0.34403977 0.0101972907
#>  [ reached 'max' / getOption("max.print") -- omitted 1667 rows ]