Read DART Basin TribPit observation files and combine with auxiliary age data

Read DART Basin TribPit observation files and combine with auxiliary age data

Usage

read_DART_file(
  filepath,
  aux_age_df,
  DART_config = "https://www.cbr.washington.edu/downloads/paramest/sites_config.txt",
  ...
)

Arguments

filepath

the path to the DART Basin TribPit observation file

aux_age_df

A data frame object in the same format as aux_age_df (see documentation for s4t_ch)

DART_config

the path to a site configuration file "site_config.txt" generated by Columbia Basin Research. If blank, defaults to www.cbr.washington.edu/downloads/paramest/sites_config.txt

...

any additional file paths of DART files to include

Value

A list with two elements. The first is the formatted capture history data (ch_df) and the second is the age auxiliary data (aux_age_df).

Details

Currently, only Dams in the Snake River are checked for transported fish.